requirements.txt
django-tailwind       3.8.0
pillow                11.0.0
pip                   24.2
sqlparse              0.5.1
tzdata                2024.2